SetUpClient and SetUpClientImpl updated
The SetUpClient interface now extends Remote and has the method
Options() which runs through the various option of the SetUpClient.

SetUpClientImpl has been updated to implement the Option() method to
enable users to loop through the various options.
Also there is now a 3rd option in Options to enable the user to QUIT
QuizMaster - once fully implemented this will close the client thread.