Method added to add a player, and bug fix in QuizGameImpl
PlayerClientImpl now has the addPlayer() method to add a player onto
the server, and returns a player ID. This method has been updated on
the QuizGame interface - @param Player changed to String userName.
QuizGameImpl updated to use a different method to return unique IDs.
This is to enable it to be made thread safe more easily at a later date.