PlayerClientImpl updated
so that the players score can be added to the quiz, and a list of the
top 5 scores for that quiz is shown.

Methods to implement this added to the interface Quiz and QuizGame.

Class ScoreImpl has been made serializable to allow objects of
ScoreImpl to be passed to the client.