Updates to add a players score and return results
Implementation of sendResult() method in QuizGameImpl which adds a
players score to the scores list, sorts the scores into order, and
returns the top 5 scores. Method getTopFiveScores() sorts the scores
for a quiz and returns the top 5 as a List of Strings.

This method is utilised in PlayerClientImpl which has the added method
getQuizToPlay() which returns a quiz by its ID number.

ScoreImpl; note method compareTo now returns -1 if this.score > score ;
to correctly order the scores from highest to lowest.